Finished Project (Not Very Accurate)
I finished the last part of the assignment,  the weighting system is just the scores for each individual comparison of one word, two words and three words phrases, and punctuation. the result is the composite score. Also made the output look similar to the example presented in Canvas and so much more...